Separated Components to Different Files
Separated components and added additional comments
